requirements.txt
greenlet==1.1.1
pika==1.2.0
psycopg2-binary==2.9.1
python-dotenv==0.19.0
six==1.16.0
SQLAlchemy==1.4.22
SQLAlchemy-Utils==0.37.8